+++ src/app/payment/page.tsx
@@ -1,9 +1,9 @@
 'use client';
-import PaymentIndex from '@/components/payment/PaymentIndex';
+import PaymentIndex from '@/app/payment/(payment)/PaymentIndex';
 import React from 'react';
 
 import { usePaymentStatus } from '@/factory/Payment/paymentcheck';
-import NotPaymentIndex from '@/components/payment/NotPaymentIndex';
+import NotPaymentIndex from '@/app/payment/(payment)/NotPaymentIndex';
 import PaymentSkeleton from '@/components/skeleton/PaymentSkeleton';
 import { IsSubscribed } from '@/utils/issubscribed';
 const PaymentPage = () => {
